requirements.txt
Flask==2.3.3
Flask-Login==0.6.3
Flask-SocketIO==5.3.6
Flask-WTF==1.1.1
Flask-Limiter==3.5.0
Werkzeug==2.3.7
python-socketio==5.9.0
eventlet==0.33.3
bcrypt==4.0.1
python-dotenv==1.0.0
gunicorn==21.2.0